Downloads/Travaux  Hackaton Groupe 8/Projet-Groupe_8/Demo_Livrable_3/requirements.txt
fastapi
uvicorn
casbin
bcrypt
python-jose
sqlalchemy
pydantic
python-multipart
jinja2
httpx
pytest
